basic-langchain/pyproject.toml
[project]
name = "exercice1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.31",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.1.1",
]
